byte fields from integer to long (issue 379)

+++ output/typescript/types.ts
@@ -10336,7 +10336,7 @@ export interface IndicesDataStreamsStatsDataStreamsStatsItem {
   data_stream: Name
   maximum_timestamp: EpochTime<UnitMillis>
   store_size?: ByteSize
-  store_size_bytes: integer
+  store_size_bytes: long
 }
 
 export interface IndicesDataStreamsStatsRequest extends RequestBase {
@@ -10350,7 +10350,7 @@ export interface IndicesDataStreamsStatsResponse {
   data_stream_count: integer
   data_streams: IndicesDataStreamsStatsDataStreamsStatsItem[]
   total_store_sizes?: ByteSize
-  total_store_size_bytes: integer
+  total_store_size_bytes: long
 }
 
 export interface IndicesDeleteRequest extends RequestBase {
